Almost done with User Management, I know I suck

+++ app/control-panel/controllers/UsersCtrl.js
@@ -10,21 +10,18 @@ function UsersCtrl($scope, $state, $stateParams, admin) {
   function loadUsers() {
     var userPromise = admin.getAllUsers();
     userPromise.then(function(response) {
-      console.log(response);
       $scope.users = []; // To empty before reloading
       var i;
       for (i = 0 ; i < response.length; i++) {
         $scope.users.push(response[i]);
       }
-
-
     }).catch(function(response) {
       addErrorAlert(response.data);
     });
   }
 
   $scope.register = function(user) {
-    console.log(user);
+
     var registerPromise = admin.registerUser(user);
     registerPromise.then(function(response) {
       addSuccessAlert('User Successfully Registered!');
@@ -32,7 +29,6 @@ function UsersCtrl($scope, $state, $stateParams, admin) {
     })
     .catch(function(response) {
       var errors = response.data.errors.full_messages;
-      // Var errors = [];
       var i;
       for (i = 0 ; i < errors.length ;i++) {
         addErrorAlert(errors[i]);
@@ -40,43 +36,35 @@ function UsersCtrl($scope, $state, $stateParams, admin) {
     });
   };
 
-  $scope.delete = function(_id) {
-    var deletePromise = admin.deleteUser(_id);
-    deletePromise.then(function() {
-      loadUsers();
-      addSuccessAlert('User Deleted!');
-    })
-    .catch(function() {
+  $scope.deleteUser = function() {
+    if ($scope.modalUserID >= 0) {
+      var deletePromise = admin.deleteUser(_id);
+      deletePromise.then(function() {
+        loadUsers();
+        addSuccessAlert('User Deleted!');
+      })
+      .catch(function() {
 
-    });
+      });
+    }
+    $scope.modalUserID = -1;
 
 
 
   };
 
-  function addSuccessAlert(alert) {
-    alerts.push({type: 'success', msg: alert});
-  }
-
-  function addErrorAlert(alert) {
-    alerts.push({type: 'danger', msg: alert});
-  }
-
-  $scope.closeAlert = function(index) {
-    alerts.splice(index, 1);
+  $scope.deleteModal = function(_id) {
+    $scope.showDeleteModal = !$scope.showDeleteModal;
+    $scope.modalUserID = _id;
   };
+
+
   alerts = [];
 
-  $scope.showModal = false;
-  $scope.toggleModal = function() {
-    $scope.showModal = !$scope.showModal;
-    console.log($scope.showModal);
-  };
-
-
+  $scope.showDeleteModal = false; // Set to true to show the delete confirmation modal
+  $scope.showEditModal = false; // Set true to show Edit Modal
+  $scope.modalUserID = -1;
   $scope.users = [];
-
-
   loadUsers();
 
 }
